--- Assets/Scripts/Player/WeaponType/Weapon/Bren Gun.asset
@@ -1,31 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dd14c3d7a5976ac479afda84b23d5661, type: 3}
-  m_Name: Bren Gun
-  m_EditorClassIdentifier: 
-  p_WeaponName: Bren Gun
-  p_WeaponModel: {fileID: 0}
-  p_WeaponFireRate: 6
-  p_WeaponChargeRate: 0
-  p_WeaponChargeCap: 0
-  p_BulletCount: 1
-  p_BulletInterval: 0
-  p_TotalAmmo: 30
-  p_ProjectileType: {fileID: 584065255677097293, guid: ca3ad3c93c23dc64187cd331fff8942f, type: 3}
-  p_BulletSpeed: 35
-  p_BulletMaxDamage: 25
-  p_BulletMinDamage: 20
-  p_TimeBeforeSelfDestruct: 3
-  p_GunShotAudio: {fileID: 0}
-  p_GunPickupAudio: {fileID: 0}
-  isAuto: 1
-  isCharge: 0
