+++ Assets/Scripts/Weapons/PlayerFist.cs
@@ -9,17 +9,35 @@ public class PlayerFist : MonoBehaviour
     public float punchCooldownMax = 1f;
     public float punchCooldownCurrent;
 
-    public int knockbackForce;
+    public GameObject player;
+    public float knockbackForce;
+    public GameObject HitShotEffect;
+
     public bool isGrabbing = false; // prop is being grabbed in hand
+    [SerializeField]
+    private Transform cameraTransform;
+    [SerializeField]
+    private LayerMask pickUpLayerMask;
+    [SerializeField]
+    private Transform objectGrabPoint;
+    public float grabDistance;
+    private PropGrab propGrab;
+    public float throwForce;
 
+    public Animator anim;
 
-    Animator anim;
+    public bool canBigFist = false;
+    public BigFistCollider bigFistCollider;
+
 
     // Start is called before the first frame update
     void Start()
     {
+        player = GameObject.FindGameObjectWithTag("Player");
         anim = GetComponent<Animator>();
         punchCooldownCurrent = punchCooldownMax;
+
+        cameraTransform = Camera.main.transform;
     }
 
     // Update is called once per frame
@@ -33,15 +51,31 @@ public class PlayerFist : MonoBehaviour
 
         if (Input.GetMouseButtonDown(0) && punchCooldownCurrent >= punchCooldownMax)
         {
-            Punch();
+            if (canBigFist == true)
+            {
+                BigFist();
+            }
+            else
+            {
+                Punch();
+            }
         }
-        if (Input.GetKeyDown(KeyCode.F) && isGrabbing == false)
+
+        if (Input.GetKeyDown(KeyCode.F))
         {
-            Grab();
+            if (propGrab == null) // not carrying an object
+            {
+                Grab();
+            }
+            else
+            {
+                Throw(throwForce);
+            }
         }
+
         if (Input.GetKeyDown(KeyCode.F) && isGrabbing == true)
         {
-            Throw();
+            Throw(throwForce);
         }
     }
 
@@ -50,38 +84,48 @@ public class PlayerFist : MonoBehaviour
         if (other.GetComponent<EnemyHealth>() != null)
         {
             other.GetComponent<EnemyHealth>().TakeDamage(damage);
+            Instantiate(HitShotEffect, other.transform.position, other.transform.rotation);
+            other.gameObject.transform.position = Vector3.MoveTowards(other.transform.position, player.transform.position, -knockbackForce);
         }
     }
 
     public void Punch()
     {
         punchCooldownCurrent = 0;
-        if (isGrabbing == true) // throw prop
+        if (propGrab != null) // carrying an object
         {
-            Throw();
+            Throw(throwForce);
         }
         else // punch (nothing grabbed in hand)
         {
             anim.SetTrigger("Punch");
         }
-        //StartCoroutine(ResetPunch());
     }
 
     public void Grab()
     {
-        isGrabbing = true;
-        anim.SetTrigger("Grab");
+        if (Physics.Raycast(cameraTransform.position, cameraTransform.forward, out RaycastHit raycastHit, grabDistance, pickUpLayerMask))
+        {
+            if (raycastHit.transform.TryGetComponent(out propGrab))
+            {
+                propGrab.Grab(objectGrabPoint);
+            }
+        }
     }
 
-    public void Throw()
+    public void Throw(float throwForce)
     {
-        isGrabbing = false;
         anim.SetTrigger("Throw");
+
+        propGrab.Throw(throwForce);
+        propGrab = null;
     }
 
-    //public IEnumerator ResetPunch()
-    //{
-    //    yield return new WaitForSeconds(0.25f);
-    //    anim.SetBool("Punch", false);
-    //}
+    public void BigFist()
+    {
+        anim.SetTrigger("BigFist");
+
+        bigFistCollider.canBigFist = true;
+        canBigFist = false;
+    }
 }
